+++ spec/dummy/app/controllers/foobars_controller.rb
@@ -3,8 +3,9 @@ class FoobarsController < ApplicationController
   include RestfulJson::Authorizing
   respond_to :json
   
+  can_filter_by :a_query, with_query: ->(t, q, value) {q.where(foo_id: value)}
   can_filter_by :foo_id
-  can_filter_by :foo_date, :bar_date, using: [:lt, :eq, :gt], with_default: Time.now
+  can_filter_by :foo_date, :bar_date, using: [:lt, :eq, :gt]
   supports_functions :count
   can_order_by :foo_id
   default_order [{foo_id: :desc}]